refactor AsyncRunner usage - remove as param to HgRunner methods; pullBundleChunk now uses for loop to avoid INPROGRESS status
The INPROGRESS status will only be sent after waiting up to 14 seconds for more data to pull as the bundle is being created.  It is preferable to simply wait on the server and send back data when it is available, as long as the wait is less than the client timeout, which is currently 20 seconds.